backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.17
pandas==2.2.3
openpyxl==3.1.5
weasyprint==63.1
httpx==0.28.1
email-validator==2.2.0
stripe==11.1.0
